DentalXrayAI/requirements.txt
ultralytics
flask
flask-cors
waitress
pillow
roboflow